requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
chromadb
streamlit
pypdf
python-dotenv